Add Internal and JavaScript categories (#24744)
Summary:
So far we have only officially supported "iOS" and "Android" as categories, with a "General" catch all for everything else.
Let's add a "JavaScript" category for changes that only affect JavaScript, and an "Internal" category for things that do not make sense in a changelog (e.g. CI changes).

[Internal] [Added] - Added JavaScript and Internal categories to changelog regex
Pull Request resolved: https://github.com/facebook/react-native/pull/24744

Differential Revision: D15258061

Pulled By: cpojer

fbshipit-source-id: 8b6cb193adf602d3bd6639b550e04775b201fe59