Contributing Guide Revamp (#24705)
Summary:
This is the first step in revamping the Contribution Guide at `/CONTRIBUTING.md`. This work will require changes to the React Native website repository, as well as the React Native GitHub wiki. For review purposes, changes to copy that exists outside of this repository will be included within a `.github/wiki` directory in this PR, but they are not intended to be part of the final PR that gets merged.

[General] [Changed] - Contribution Guide Revamp
Pull Request resolved: https://github.com/facebook/react-native/pull/24705

Differential Revision: D15290268

Pulled By: hramos

fbshipit-source-id: b09c86738e75143fbdf2fea97ccd860da79362d2